refactored to individual components for each page
