+++ index.html
@@ -56,7 +56,10 @@
   </nav>
 
   <div class="container">
-    <section id="partisan-app"></section>
+    <img class="center-block" src="./assets/images/HeaderFactoryLogo.jpg" />
+    <section>
+        <div id="partisan-app" />
+    </section>
   </div>
   <!-- <script src="./scripts/application.min.js"></script> -->
 
@@ -66,8 +69,13 @@
   <script type="text/javascript " src="scripts/utils.js "></script>
   <script type="text/javascript " src="scripts/models/index.js"></script>
   <script type="text/javascript " src="scripts/footer.js "></script>
-  <script type="text/javascript " src="scripts/app.js "></script>
-  <script type="text/javascript " src="scripts/howtogetinvolved.js "></script>
+  <script type="text/javascript " src="scripts/components/whatispartisan.js "></script>
+  <script type="text/javascript " src="scripts/components/howtogetinvolved.js "></script>
+  <script type="text/javascript " src="scripts/components/cafebar.js "></script>
+  <script type="text/javascript " src="scripts/components/howareweorganised.js "></script>
+  <script type="text/javascript " src="scripts/components/inspiration.js "></script>
+  <script type="text/javascript " src="scripts/components/contact.js "></script>
+
   <script type="text/javascript " src="scripts/application.js "></script>
 </body>
 <!-- <footer class="footer">
@@ -76,4 +84,5 @@
     <a href="https://facebook.com/benjamin.t.thomson"><strong>Click here for facebook</strong></a>
   </div>
 </footer> -->
+
 </html>
