+++ scripts/components/cafebar.js
@@ -1,0 +1,22 @@
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var Partisan;
+(function (Partisan) {
+    var Components;
+    (function (Components) {
+        var CafeBar = (function (_super) {
+            __extends(CafeBar, _super);
+            function CafeBar() {
+                _super.call(this);
+            }
+            CafeBar.prototype.render = function () {
+                return (React.createElement("div", null, React.createElement("h2", null, "Café-bar"), React.createElement("div", {className: "row"}, React.createElement("div", {className: "col-xs-12"}, React.createElement("p", null, "The space will include a cafe-bar area with regular opening hours. The cafe-bar will sell affordable alcoholic and non-alcoholic drinks, snacks and possibly hot food. It will also sell a selection of books from independent publishers (e.g. PM Press, Verso, Pluto etc) and house a Feminist Library (contributed by Feminist Fightback)."), React.createElement("p", null, "Although the space will not require that people within it consume food or drinks, the space as a whole will be supported by funds raised through the café-bar. So eating/drinking in it will be a way of showing solidarity and support and keeping the space running."), React.createElement("p", null, "The café-bar will be collectively and democratically owned and managed as a workers co-op."))), React.createElement("h2", null, "Socal space"), React.createElement("div", {className: "row"}, React.createElement("div", {className: "col-xs-12"}, React.createElement("p", null, "The social space will include a flexible venue room that is separate to the cafe-bar. This space can be used by different individuals and groups for club nights and gigs, film screenings, fund raisers and meetings. The space will be freely available to non-profit and social justice groups, such as anti-cuts campaigners, migrant groups, queer collectives and others."), React.createElement("p", null, "Priority for the space will be given to community and political groups and arts organisations organized as collectives, with a commitment to progressive values.")))));
+            };
+            return CafeBar;
+        }(React.Component));
+        Components.CafeBar = CafeBar;
+    })(Components = Partisan.Components || (Partisan.Components = {}));
+})(Partisan || (Partisan = {}));
