--- scripts/howtogetinvolved.js
@@ -1,22 +1,0 @@
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
-var Partisan;
-(function (Partisan) {
-    var Components;
-    (function (Components) {
-        var HowToGetInvolved = (function (_super) {
-            __extends(HowToGetInvolved, _super);
-            function HowToGetInvolved() {
-                _super.call(this);
-            }
-            HowToGetInvolved.prototype.render = function () {
-                return (React.createElement("div", null, React.createElement("img", {className: "center-block", src: "./assets/images/HeaderFactoryLogo.jpg"}), React.createElement("div", null, React.createElement("h2", null, "How to get  involved  help"), React.createElement("div", {className: "row"}, React.createElement("div", {className: "col-xs-12"}, React.createElement("p", null, "What we really need right now is people. We need your energy, enthusiasm and ideas to get this off the ground."), React.createElement("p", null, "Partisan will only work with your help, support and involvement. So please do get involved in any way you can."), React.createElement("h2", null, "Are you an individual who wants to join the collective?"), React.createElement("p", null, "Do you want to help set up the venue? Work in it? Do you have building or design skills? Do you have fundraising skills? Do you just want to help establish an alternative venue for culture and politics in Manchester? Let us ", React.createElement("a", {href: "#contact"}, "know"), "."), React.createElement("h2", null, "Are you a community group or political organisation?"), React.createElement("p", null, "Get in touch to tell us that you would like to become a User Group and use the space then ", React.createElement("a", {href: "#contact"}, "contact us"), "."))))));
-            };
-            return HowToGetInvolved;
-        }(React.Component));
-        Components.HowToGetInvolved = HowToGetInvolved;
-    })(Components = Partisan.Components || (Partisan.Components = {}));
-})(Partisan || (Partisan = {}));
